Remove need to register frame testers
Allow frame testers to be directly accessed, removing the need to register them and reference via ResourceLocation.
